<feat>:<Ajout des checkbox est coach et de la gestion des images> <fix>:<Correction des deux classes css et de l'emplacement d'Autocomplete>
